requirements.txt
-c https://raw.githubusercontent.com/apache/airflow/constraints-2.10.3/constraints-3.11.txt

apache-airflow[amazon,common-sql,microsoft-mssql,sftp]==2.10.3
paramiko==3.5.0
python-gnupg==0.5.4
requests==2.32.3
pandas